pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "echo-api"
version = "1.0.0"
description = "Echo — dual-feed 18+ social platform API (Austin, TX / 2026 Compliance)"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",

    # Database — async only (Rule 4)
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",

    # Validation (Rule 6 — Pydantic v2 mandatory)
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.2.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # AWS S3
    "boto3>=1.35.0",
    "aiofiles>=24.1.0",

    # Redis (async)
    "redis[hiredis]>=5.2.0",

    # HTTP client (async — Rule 4)
    "httpx>=0.27.0",

    # File uploads
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    "fakeredis>=2.26.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pip-audit>=2.7.0",
    "types-passlib",
    "types-boto3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=80"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "S", "ASYNC"]
ignore = ["S101"]   # allow assert in tests

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.coverage.run]
omit = ["tests/*", "alembic/*"]
